Extract AppTabs into its own file.
