+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_code_challenge/pages/app_tabs.dart';
 import 'package:flutter_code_challenge/pages/auth/home/auth_home_page.dart';
 import 'package:flutter_code_challenge/pages/auth/weather/weather_page.dart';
 import 'package:flutter_code_challenge/pages/home/home_page.dart';
@@ -8,6 +9,7 @@ import 'package:provider/provider.dart';
 
 void main() {
   runApp(
+    // Register our providers from the top-level.
     MultiProvider(
       providers: [
         ChangeNotifierProvider(
@@ -25,51 +27,16 @@ void main() {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    context.read<Auth>().init();
+
     return MaterialApp(
       title: 'Flutter Code Challenge',
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
+      // Switch between page sets here. If logged-in, display AppTabs...
+      // Otherwise, display HomePage.
       home: context.watch<Auth>().isLoggedIn ? AppTabs() : HomePage(),
-    );
-  }
-}
-
-class AppTabs extends StatefulWidget {
-  final List<Widget> pages = [
-    AuthHomePage(),
-    WeatherPage(),
-  ];
-
-  @override
-  _AppTabsState createState() => _AppTabsState();
-}
-
-class _AppTabsState extends State<AppTabs> {
-  int currentIndex = 0;
-
-  void itemTapped(int index) {
-    setState(() => currentIndex = index);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: currentIndex,
-        items: [
-          BottomNavigationBarItem(
-            label: 'Home',
-            icon: Icon(Icons.home),
-          ),
-          BottomNavigationBarItem(
-            label: 'Weather',
-            icon: Icon(Icons.wb_sunny),
-          ),
-        ],
-        onTap: itemTapped,
-      ),
-      body: widget.pages[currentIndex],
     );
   }
 }
